added uninstall info and colors :-)

+++ install.sh
@@ -51,30 +51,65 @@ mongo_download="$mongo_download$mongo_tar"
 work_fld="/var/tmp/mongo"
 
 # Folder where Mongo will be installed locally
-tar_install_fld="/usr/local/lib"
+install_fld="/usr/local/lib"
+
+# Bin folder in which mongodb symlinks will be created
+bin_folder="/usr/local/bin"
 
 orig_fld="$PWD"
+
+#--------------------------------------------
+# color codes
+
+C_BLACK='\033[0;30m'
+C_RED='\033[1;31m'
+C_GREEN='\033[1;32m'
+C_YELLOW='\033[1;33m'
+C_BLUE='\033[1;34m'
+C_MAGENTA='\033[1;35m'
+C_CYAN='\033[1;36m'
+C_WHITE='\033[1;37m'
+C_GRAY='\033[;37m'
+C_DEFAULT='\E[0m'
+
+#-------------------------------------------
+
+function cecho ()  {
+  # Description
+  #   Colorized echo
+  #
+  # Parameters:
+  #   $1 = message
+  #   $2 = color
+  
+  message=${1:-""}   # Defaults to default message.
+  color=${2:-$C_BLACK}           # Defaults to black, if not specified.
+  
+  echo -e "$color$message$C_DEFAULT"  
+  
+  return
+} 
 
 # ----------------------------------------
 
 # this script should be run under ROOT account
 if [ "$UID" -ne "0" ]; then
-  echo "Error: Script should be run under root user privileges."
+  cecho "Error: Script should be run under root user privileges." $C_RED
   exit $ERR_START
 fi 
 
 if [ ! "`which sed`" ]; then
-  echo "Error: 'sed' not found. Please, install 'sed' first."
+  cecho "Error: 'sed' not found. Please, install 'sed' first." $C_RED
   exit $ERR_START
 fi
 
 if [ ! "`which sudo`" ]; then
-  echo "Error: 'sudo' not found. Please, install 'sudo' first."
+  cecho "Error: 'sudo' not found. Please, install 'sudo' first." $C_RED
   exit $ERR_START
 fi
 
 if [ ! "`which wget`" ]; then
-  echo "Error: 'wget' not found. Please, install 'wget' first."
+  cecho "Error: 'wget' not found. Please, install 'wget' first." $C_RED
   exit $ERR_START
 fi
 
@@ -86,7 +121,7 @@ fi
 # test if we are able to write to work folder
 sudo touch $work_fld/test
 if [ $? -ne 0 ]; then
-  echo "Fatal error: can't open work folder $work_fld"
+  cecho "Fatal error: can't open work folder $work_fld" $C_RED
   exit $ERR_PREPARE
 fi
 
@@ -98,32 +133,35 @@ cd $work_fld
 # download archive
 sudo wget $mongo_download
 if [ $? -ne 0 ]; then
-  echo "Fatal error: can not download $mongo_download."
+  cecho "Fatal error: can not download $mongo_download." $C_RED
   exit $ERR_PREPARE
 fi
 
 # extract original source archive to predefined folder
-sudo tar xvzf $mongo_tar -C $tar_install_fld
+sudo tar xvzf $mongo_tar -C $install_fld
 if [ $? -ne 0 ]; then
-  echo "Fatal error: can not extract archive $mongo_tar to folder $tar_install_fld."
+  cecho "Fatal error: can not extract archive $mongo_tar to folder $install_fld." $C_RED
   exit $ERR_PREPARE
 fi
 
-if [ ! -d $tar_install_fld/${mongo_tar%.*} ]; then
-  echo "Fatal error: expected folder does not exist $tar_install_fld/${mongo_tar%.*} ."
+if [ ! -d $install_fld/${mongo_tar%.*} ]; then
+  cecho "Fatal error: expected folder does not exist $install_fld/${mongo_tar%.*} ." $C_RED
   exit $ERR_PREPARE
 fi
 
-if [ ! -e "$tar_install_fld/${mongo_tar%.*}/bin/mongod" ]; then
-  echo "Fatal error: mongod not found in $tar_install_fld/${mongo_tar%.*}."
+if [ ! -e "$install_fld/${mongo_tar%.*}/bin/mongod" ]; then
+  cecho "Fatal error: mongod not found in $install_fld/${mongo_tar%.*}." $_CRED
   exit $ERR_PREPARE
 fi
 
 ############################### Files should be already there ###############
 
+# Create uninstall folder
+[ -d $install_fld/.uninstall ] || sudo mkdir $install_fld/.uninstall
+
 # Create system user for MongoDB
 username="mongodb"
-echo "Creating system user '"$username"'" 
+cecho "Creating system user '"$username"'" $C_GREEN
   
 # Check if user does not exist already
 if [ ! "`sed -n -e "/^$username:/p" /etc/passwd`" ]
@@ -132,44 +170,60 @@ then
   sudo useradd --create-home $username
   if [ "$?" -ne "0" ]
   then 
-    echo "Fatal error: User '"$username"' creation failed."
+    cecho "Fatal error: User '"$username"' creation failed." $C_RED
     exit $ERR_INSTALL
   else
     echo "User '"$username"' created."
+	
+	# and make note that user has been created
+	sudo touch $install_fld/.uninstall/user_created
   fi
 else
   echo "User '"$username"' already does exist."
 fi 
 
-echo "Creating symlinks in /usr/local/bin..."
-sudo cp -s  $tar_install_fld/${mongo_tar%.*}/bin/* /usr/local/bin
+cecho "Creating symlinks in $bin_folder..." $C_GREEN
+# move existing files/symlinks to .uninstall folder
+[ -d $install_fld/.uninstall/bin ] || sudo mkdir $install_fld/.uninstall/bin
+for f in $install_fld/${mongo_tar%.*}/bin/*; do
+  sudo mv $bin_folder/${f##/*/} $install_fld/.uninstall/bin > /dev/nul
+done
+
+# create new symlinks pointing to new folder
+sudo cp -s  $install_fld/${mongo_tar%.*}/bin/* $bin_folder
 if [ $? -ne 0 ]; then
-  echo "Fatal Error: Creation of symlinks failed."
+  cecho "Fatal Error: Creation of symlinks failed." $C_RED
   exit $ERR_INSTALL
 fi
 
-echo "Creating database folder '"$mongo_dbpath"'"
-[ ! -d $mongo_dbpath ] && sudo mkdir $mongo_dbpath
+# create folder for database files
+cecho "Creating database folder '"$mongo_dbpath"'" $C_GREEN
+[ -d $mongo_dbpath ] || sudo mkdir $mongo_dbpath
 if [ $? -ne 0 ]; then
-  echo "Fatal Error: Not possible to create folder for database."
+  cecho "Fatal Error: Not possible to create folder for database." $C_RED
   exit $ERR_INSTALL
 fi
 
+# change owner for database folder so mongodb daemons can access it
 sudo chown $username:$username $mongo_dbpath
 
-echo "Copying additional configuration files necessary for MongoDB."
+cecho "Copying additional configuration files necessary for MongoDB." $C_GREEN
 
-# create necessary folders
-[ -d "/var/log/mongodb" ] && sudo rm -R /var/log/mongodb
+# create folder for log files
 [ -d "/var/log/mongodb" ] || sudo mkdir /var/log/mongodb
 
+# and change log folder owner so mongodb daemons can access it
 sudo chown $username:$username /var/log/mongodb
 
-# create init.d script
 echo "Creating init.d script for MongoDB database server..."
+# move existing init.d script to unistall folder
+[ -d $install_fld/.uninstall/initd ] || sudo mkdir $install_fld/.uninstall/initd
+sudo mv $orig_fld/etc/init.d/mongod $install_fld/.uninstall/initd > /dev/nul
+
+# create new init.d script
 sudo cp $orig_fld/etc/init.d/mongod /etc/init.d/mongod
 if [ $? -ne 0 ]; then
-  echo "Fatal Error: MongoDB database server init.d script copying failed."
+  cecho "Fatal Error: MongoDB database server init.d script creation failed." $C_RED
   exit $ERR_INSTALL
 fi
 
@@ -177,29 +231,31 @@ fi
 sudo chmod +x /etc/init.d/mongod
 
 # add configuration file
-echo "Copying base configuration file..."
+cecho "Copying base configuration file (keeping existing ones)..." $C_GREEN
 [ -d /etc/mongodb ] || sudo mkdir /etc/mongodb
-sudo cp -r $orig_fld/etc/mongodb/* /etc/mongodb
-if [ $? -ne 0 ]; then
-  echo "Fatal Error: Copying of configuration files failed."
-  exit $ERR_INSTALL
-fi
+for f in $orig_fld/etc/mongodb/*; do
+  [ -e /etc/mongodb/${f##/*/} || sudo cp $f /etc/mongodb
+  if [ $? -ne 0 ]; then
+    cecho "Error: Copying of configuration file $f failed." $C_RED
+    exit $ERR_INSTALL
+  fi
+done
 
 # configure log-rotate
-echo "Configuring logrotate..."
-sudo cp $orig_fld/etc/logrotate.d/mongodb /etc/logrotate.d
+cecho "Configuring logrotate..." $C_GREEN
+[ -e /etc/logrotate.d/mongodb ] || sudo cp $orig_fld/etc/logrotate.d/mongodb /etc/logrotate.d
 if [ $? -ne 0 ]; then
-  echo "Fatal Error: Copying of logrotate configuration files failed."
+  cecho "Error: Copying of logrotate configuration files failed." $C_RED
   exit $ERR_INSTALL
 fi
 
 cd $orig_fld
 
 # make MongoDB to start at server boot
-echo "Creating startup scripts..."
+cecho "Creating startup scripts..." $C_GREEN
 sudo update-rc.d mongod defaults 80 20
 if [ $? -ne 0 ]; then
-  echo "Fatal Error: Not able to create startup rc scripts for MongoDB."
+  cecho "Fatal Error: Not able to create startup rc scripts for MongoDB." $C_RED
   exit $ERR_INSTALL
 fi
